+++ release-notes.md
@@ -7,6 +7,7 @@
 - Allow X windows to scale themselves.
 - Implement ext-image-capture-source-v1.
 - Implement ext-image-copy-capture-v1.
+- Implement screencast session restoration.
 
 # 1.6.0 (2024-09-25)